nass-1260: v nice refactor that removes need to useMateralizedViewRefreshStatsQuery
