+++ packages/web/app/components/Table/RefreshStatsDisplay.jsx
@@ -1,9 +1,12 @@
 import { Box } from '@material-ui/core';
-import React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { BodyText } from '../Typography';
 import { TranslatedText } from '../Translation';
 import styled from 'styled-components';
 import { Colors } from '../../constants';
+import { useParsedCronExpression } from '../../utils/useParsedCronExpression';
+import { formatDistanceToNow, parseISO } from 'date-fns';
+import { useTranslation } from '../../contexts/Translation';
 
 const SmallText = styled(BodyText)`
   font-size: 10px;
@@ -23,9 +26,37 @@ const StatsError = () => (
   </ErrorText>
 );
 
-export const RefreshStatsDisplay = ({ stats, error }) => {
-  if (!stats) return null;
-  const { lastUpdated, schedule } = stats;
+export const RefreshStatsDisplay = ({
+  stats,
+  error,
+  recalculateDistanceFromNowIntervalMs = 1000 * 60,
+}) => {
+  const { getTranslation } = useTranslation();
+  const { schedule, lastRefreshed } = stats;
+  const parsedSchedule = useParsedCronExpression(schedule);
+  const [lastUpdated, setLastUpdated] = useState();
+
+  const dateAsDistanceToNow = useCallback(
+    date =>
+      formatDistanceToNow(parseISO(date), {
+        addSuffix: getTranslation('schedule.distanceFromNow.suffix', 'ago'),
+      }),
+    [getTranslation],
+  );
+
+  const handleRefreshLastUpdated = useCallback(() => {
+    setLastUpdated(dateAsDistanceToNow(lastRefreshed));
+  }, [dateAsDistanceToNow, lastRefreshed]);
+
+  // Update the distance from now text every minute
+  useEffect(() => {
+    if (!lastRefreshed) return;
+    const interval = setInterval(handleRefreshLastUpdated, recalculateDistanceFromNowIntervalMs);
+    handleRefreshLastUpdated();
+    return () => clearInterval(interval);
+  }, [handleRefreshLastUpdated, lastRefreshed, recalculateDistanceFromNowIntervalMs]);
+
+  if (!lastUpdated) return null;
   return (
     <Box display="flex" flexDirection="column" alignItems="flex-end">
       {error ? (
@@ -43,7 +74,7 @@ export const RefreshStatsDisplay = ({ stats, error }) => {
             <TranslatedText
               stringId="table.refreshSchedule.schedule"
               fallback="Updated :schedule"
-              replacements={{ schedule: schedule.toLowerCase() }}
+              replacements={{ schedule: parsedSchedule.toLowerCase() }}
             />
           </SoftText>
         </>
